requirements.txt
psutil==5.2.2
azure_common==1.1.6
